Splitt urelaterte tester til to filer
