+++ src/main/kotlin/no/nav/helse/flex/soknadsopprettelse/SoknadGenerering.kt
@@ -21,9 +21,11 @@ fun erForsteSoknadIForlop(
     sykepengesoknad: Sykepengesoknad
 ): Boolean {
     return eksisterendeSoknader.asSequence()
-        // Finner søknader med samme arbeidssituasjon som, men med 'fom' før eller lik søknaden det sammenlignes med.
+        // Finner søknader med samme arbeidssituasjon som, men med 'fom' FØR eller LIK søknaden det sammenlignes med.
         .filter { it.fom != null && it.fom.isBeforeOrEqual(sykepengesoknad.fom!!) }
         .filter { it.sykmeldingId != null && it.startSykeforlop != null }
+        // Spørsmål om medlemskap vil bare bli stilt i for søknader med arbeidssituasjon.ARBEIDSTAKER, men det ingen
+        // gevinst i å eksplitt sjekke på det her.
         .filter { it.arbeidssituasjon == sykepengesoknad.arbeidssituasjon }
         // Sjekker om det finnes en tidligere søknad med samme startdato for sykeforløp.
         .none { it.startSykeforlop == sykepengesoknad.startSykeforlop }
@@ -41,12 +43,10 @@ fun harBlittStiltUtlandsSporsmal(
         .any { it -> it.sporsmal.any { it.tag == UTENLANDSK_SYKMELDING_BOSTED } }
 }
 
-// Finner søknader med samme arbeidssituasjon som, men med 'fom' før søknaden det sammenlignes med.
+// Finner søknader med samme arbeidssituasjon som, men med 'fom' FØR søknaden det sammenlignes med.
 private fun Sequence<Sykepengesoknad>.finnTidligereSoknaderMedSammeArbeidssituasjon(sykepengesoknad: Sykepengesoknad): Sequence<Sykepengesoknad> =
     this.filter { it.fom != null && it.fom.isBefore(sykepengesoknad.fom) }
         .filter { it.sykmeldingId != null && it.startSykeforlop != null }
-        // Spørsmål om medlemskap vil bare bli stilt i for søknader med arbeidssituasjon ARBEIDSTAKER, men det ingen
-        // gevinst i å eksplitt sjekke på det her.
         .filter { it.arbeidssituasjon == sykepengesoknad.arbeidssituasjon }
 
 // Om det er arbeidssituasjon ARBEIDSTAKER med søknadstype BEHANDLINGSDAGER, GRADERT_REISETILSKUD eller ARBEIDSTAKER
